fix: TSteps interface
`type TSteps` stores general information about how the steps are defined
Page state is stored in variable `isStepsFinished` for performance
Otherwise, the whole `steps` variable will be updated on each render

With this idea, status should not be here(next commit).
`finished` attr is more like actionOnFinished and is not used.
The order of the steps is controlled by the attr `stepId`.
`stepId` is not perfect, but we use it to use the interface of `Stepper`